Make 2 tests work
